Filter repeated lines

+++ STT/load_transcript.py
@@ -31,16 +31,17 @@ def loadTranscript(arglist=None):
                               fromfile_prefix_chars='@')
 
     parser.add_argument('-v', '--verbosity',  type=int, default=1, private=True)
-    
-    parser.add_argument('-c', '--csvfile',    type=str, 
+
+    parser.add_argument('-c', '--csvfile',    type=str,
                                               help='Output CSV file', output=True)
     parser.add_argument('-d', '--database',   type=str,
                                               help='SQLAlchemy database specification')
-    
+
     parser.add_argument('--no-comments',      action='store_true', help='Do not output descriptive comments')
 
     parser.add_argument('--logfile',   type=str, help="Logfile", private=True)
     parser.add_argument('--no-logfile', action='store_true', help='Do not output a logfile')
+    parser.add_argument('--no-header',  action='store_true', help='Do not output header to CSV file')
 
     parser.add_argument('infile',             type=str, help="Filename pattern to match", input=True)
 
@@ -58,7 +59,7 @@ def loadTranscript(arglist=None):
         csvwriter=csv.DictWriter(csvfile, fieldnames=['name', 'channel', 'datetime', 'text'])
         if not args.no_header:
             csvwriter.writeheader()
-        
+
     if args.database:
         if not args.no_logfile:
             if args.logfile:
@@ -88,7 +89,7 @@ def loadTranscript(arglist=None):
 
     filenameregexp = re.compile(R"(?P<name>.+?)(-Chan(?P<channel>[0-9]+))?-(?P<year>[0-9]{2,4})(?P<month>[0-9]{2})(?P<day>[0-9]{2})-(?P<hour>[0-9]{2})(?P<minute>[0-9]{2})(?P<second>[0-9]{2}).+", re.UNICODE)
     lineregexp = re.compile(r"^\[(?P<minute>[0-9]{2}):(?P<second>[0-9]{2}).(?P<csec>[0-9]{2})\]\s*(?P<text>.*)$", re.UNICODE)
-    
+
     for filename in glob.glob(args.infile):
         basename = os.path.basename(filename)
         filenamematch = filenameregexp.match(basename)
@@ -107,9 +108,10 @@ def loadTranscript(arglist=None):
             basetime=datetime.datetime(year, month, day, hour, minute, second)
         else:
             print("ERROR: Filename " + basename + " does not match pattern", file=sys.stderr)
-        
+
         infile = open(filename, 'r')
 
+        lasttext = ''
         for line in infile:
             linematch = lineregexp.match(line)
             if linematch:
@@ -117,26 +119,28 @@ def loadTranscript(arglist=None):
                 second = int(linematch.group('second'))
                 csec   = int(linematch.group('csec'))
                 text   =     linematch.group('text')
-                
-                finaltime = basetime + datetime.timedelta(minutes=minute, seconds=second, milliseconds = csec*10)
-                
-                if args.csvfile:
-                    csvwriter.writerow({'name':name, 'channel':channel, 'datetime':finaltime, 'text':text})
-                if args.database:
-                    connection.execute(transcript.delete().where(transcript.c.DateTime == finaltime))
-                    connection.execute(transcript.insert().values({
-                        'Name': name,
-                        'Channel': channel,
-                        'DateTime': finaltime,
-                        'Text': text}))
-                    
-    
+
+                if text != lasttext:
+                    lasttext = text
+
+                    finaltime = basetime + datetime.timedelta(minutes=minute, seconds=second, milliseconds = csec*10)
+
+                    if args.csvfile:
+                        csvwriter.writerow({'name':name, 'channel':channel, 'datetime':finaltime, 'text':text})
+                    if args.database:
+                        connection.execute(transcript.delete().where(transcript.c.DateTime == finaltime))
+                        connection.execute(transcript.insert().values({
+                            'Name': name,
+                            'Channel': channel,
+                            'DateTime': finaltime,
+                            'Text': text}))
+
     if args.csvfile:
         csvfile.close()
     if args.database:
         transaction.commit()
         connection.close()
         database.dispose()
-        
+
 if __name__ == '__main__':
     loadTranscript(None)
